+++ Program.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Data.Sqlite;
 using LabManager.Database;
+using LabManager.Repositories;
 
 var databaseSetup = new DatabaseSetup();
 
@@ -8,22 +9,17 @@ var modelAction = args[1];
 
 if (modelName == "Computer")
 {
+    var computerRepository = new ComputerRepository();
     if(modelAction == "List")
     {
         Console.WriteLine("Computer List");
-        var connection = new SqliteConnection("Data Source=database.db");
-        connection.Open();
 
-        var command = connection.CreateCommand();
-        command.CommandText = "SELECT * FROM Computers;"; 
-            
-        var reader = command.ExecuteReader();
+        var computers = computerRepository.GetAll();
 
-        while (reader.Read())
+        foreach (var computer in computers)
         {
-            Console.WriteLine("{0}, {1}, {2}", reader.GetInt32(0), reader.GetString(1), reader.GetString(2)); 
+            Console.WriteLine($"{computer.Id}, {computer.Ram}, {computer.Processor}"); 
         }
-        connection.Close();
     }
 
     if(modelAction == "New")
